adding alerts

+++ src/App.jsx
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import { useEffect, useState } from 'react'
+import Swal from 'sweetalert2';
 import './App.css'
 import UsersForm from './Components/UsersForm';
 import UsersList from './Components/UsersList';
@@ -21,8 +22,22 @@ function App() {
   }
 
   const deleteUser = (user) => {
-    axios.delete(`https://users-crud1.herokuapp.com/users/${user.id}/`)
-      .then(() => getUsers());
+    Swal.fire({
+      title: 'Do you want to delete this user?',
+      showDenyButton: true,
+      showCancelButton: true,
+      confirmButtonText: 'Yes',
+      denyButtonText: `Don't delete`,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete(`https://users-crud1.herokuapp.com/users/${user.id}/`)
+          .then(() => getUsers())
+          .catch(error => console.log(error.response?.data))
+          Swal.fire('Deleted!', '', 'success')
+      } else if (result.isDenied) {
+        Swal.fire('Cancelled', '', 'info')
+      }
+    });
   }
 
   const selectUser = (user) => setUserSelected(user);
